backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
openai==1.14.1
python-dotenv==1.0.1
requests==2.31.0